updated front end to handle pre-calculated points

+++ nfl_challenge_create.sql
@@ -78,10 +78,10 @@ CREATE TABLE "player_list" (
 	"fg404" integer,
 	"fg504" integer,
 	"xtpm4" integer,
-	"points1" integer,
-	"points2" integer,
-	"points3" integer,
-	"points4" integer,
+	"points1" decimal,
+	"points2" decimal,
+	"points3" decimal,
+	"points4" decimal,
 	CONSTRAINT "player_list_pk" PRIMARY KEY ("player_id")
 ) WITH (
   OIDS=FALSE
@@ -116,10 +116,10 @@ CREATE TABLE "def_list" (
 	"sfty4" integer,
 	"td4" integer,
 	"pts_allowed4" integer,
-	"points1" integer,
-	"points2" integer,
-	"points3" integer,
-	"points4" integer,
+	"points1" decimal,
+	"points2" decimal,
+	"points3" decimal,
+	"points4" decimal,
 	CONSTRAINT "def_list_pk" PRIMARY KEY ("def_id")
 ) WITH (
   OIDS=FALSE
